add batch mode
